+++ src/models/right_view/vae/vae_components.py
